+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- #v1 Add dummy `orbax.checkpoint` file to help identify checkpoints.
+
 ## [0.11.18] - 2025-07-01
 
 ### Changed